1stデプロイ+テスト完了

+++ db/migrate/20250523011557_change_group_id_default_in_posts.rb
@@ -1,5 +1,5 @@
 class ChangeGroupIdDefaultInPosts < ActiveRecord::Migration[6.1]
   def change
-    change_column :posts, :group_id, :integer, null: true
+    change_column :posts, :group_id, :bigint, null: true
   end
 end
